Add an alias for the --define flag to ensure the api is backwards compatible. (Nick) R=robert DELTA=9  (6 added, 0 deleted, 3 changed)
Change the default pretty printer so that then clauses of if
statements and bodies of for statements get printed in blocks
and on separate lines even if the the body is a single statement:
if (a) {
b
}
rather than
if (a) b;
This makes it easier to put breakpoints on the then clause rather
than only on the entire if statement. (Robert)
R=john
DELTA=125  (122 added, 0 deleted, 3 changed)

Revision created by MOE tool push_codebase.
MOE_MIGRATION=


git-svn-id: http://closure-compiler.googlecode.com/svn/trunk@137 b0f006be-c8cd-11de-a2e8-8d36a3108c74